+++ DHHardenSystem/Public/Invoke-LocalGPO.ps1
@@ -17,16 +17,19 @@ function Invoke-LocalGPO {
 
     .NOTES
     Name         - Invoke-LocalGPO
-    Version      - 1.7
+    Version      - 1.8
     Author       - Darren Hollinrake
     Date Created - 2021-07-24
-    Date Updated - 2025-02-15
+    Date Updated - 2025-04-13
 
     .LINK
     https://public.cyber.mil/stigs/gpo/
 
     .PARAMETER AcrobatProDC
-    DISA STIG (v2r1) - Configured Adobe Acrobat Pro (DC) in alignment with the corresponding DISA STIG. This applies both User and Computer settings.
+    DISA STIG (v2r1) - Configures Adobe Acrobat Pro (DC) in alignment with the corresponding DISA STIG. This applies both User and Computer settings.
+
+    .PARAMETER AcrobatReaderDC
+    DISA STIG (v2r1) - Configured Adobe Acrobat Reader (DC) in alignment with the corresponding DISA STIG. This applies both User and Computer settings.
 
     .PARAMETER AppLocker
     Custom - Configures AppLocker with a custom policy that allows users to run any Microsoft-signed programs AND any programs in the Program Files directories. Administrators can run anything. Valid values are 'Audit' and 'Enforce'.
@@ -85,7 +88,7 @@ function Invoke-LocalGPO {
     .PARAMETER OS
     DISA GPO - Configures the OS using the specified OS STIG. Valid values are 'Win10', 'Win11', 'Server2016', 'Server2019', and 'Server2022'. This applies both User and Computer settings.
         Windows 10 - v3r3
-        Windows 11 - v2r2
+        Windows 11 - v2r5
         Server 2016 - v2r9
         Server 2019 - v3r2 (Computer Settings Only)
         Server 2022 - v2r3 (Computer Settings Only)
@@ -119,6 +122,8 @@ function Invoke-LocalGPO {
     param (
         [Parameter(ValueFromPipelineByPropertyName)]
         [switch]$AcrobatProDC,
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [switch]$AcrobatReaderDC,
         [Parameter(ValueFromPipelineByPropertyName)]
         [ValidateSet('Audit', 'Enforce')]
         [string]$AppLocker,
@@ -177,8 +182,15 @@ function Invoke-LocalGPO {
         AcrobatProDC {
             if ($PSCmdlet.ShouldProcess("AcrobatProDC: $AcrobatProDC", "Apply GPO")) {
                 Write-LogEntry -Tee:$Tee "Applying GPO: Adobe Acrobat Pro DC"
-                & LGPO.exe /p "$CustomGPOPath\Computer - STIG - DoD Adobe Acrobat Pro DC Continuous v2r1.PolicyRules" /v >> "$($env:COMPUTERNAME)_LGPO.log"
-                & LGPO.exe /p "$CustomGPOPath\User - STIG - DoD Adobe Acrobat Pro DC Continuous v2r1.PolicyRules" /v >> "$($env:COMPUTERNAME)_LGPO.log"
+                & LGPO.exe /p "$DoDGPOPath\Computer - STIG - DoD Adobe Acrobat Pro DC Continuous v2r1.PolicyRules" /v >> "$($env:COMPUTERNAME)_LGPO.log"
+                & LGPO.exe /p "$DoDGPOPath\User - STIG - DoD Adobe Acrobat Pro DC Continuous v2r1.PolicyRules" /v >> "$($env:COMPUTERNAME)_LGPO.log"
+            }
+        }
+        AcrobatReaderDC {
+            if ($PSCmdlet.ShouldProcess("ReaderDC: $ReaderDC", "Apply GPO")) {
+                Write-Verbose "Applying GPO: Adobe Reader DC"
+                & LGPO.exe /p "$DoDGPOPath\Computer - STIG - DoD Adobe Acrobat Reader DC Continuous v2r1.PolicyRules" /v >> "$($env:COMPUTERNAME)_LGPO.log"
+                & LGPO.exe /p "$DoDGPOPath\User - STIG - DoD Adobe Acrobat Reader DC Continuous v2r1.PolicyRules" /v >> "$($env:COMPUTERNAME)_LGPO.log"
             }
         }
         Applocker {
@@ -291,13 +303,6 @@ function Invoke-LocalGPO {
                         & LGPO.exe /p "$DoDGPOPath\User - STIG - DoD Office 2019_365 v3r2.PolicyRules" /v >> "$($env:COMPUTERNAME)_LGPO.log"
                     }
                 }
-            }
-        }
-        ReaderDC {
-            if ($PSCmdlet.ShouldProcess("ReaderDC: $ReaderDC", "Apply GPO")) {
-                Write-Verbose "Applying GPO: Adobe Reader DC"
-                & LGPO.exe /p "$CustomGPOPath\Computer - STIG - DoD Adobe Acrobat Reader DC Continuous v2r1.PolicyRules" /v >> "$($env:COMPUTERNAME)_LGPO.log"
-                & LGPO.exe /p "$CustomGPOPath\User - STIG - DoD Adobe Acrobat Reader DC Continuous v2r1.PolicyRules" /v >> "$($env:COMPUTERNAME)_LGPO.log"
             }
         }
         RequireCtrlAltDel {