+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Change Log - DHHardenSystem
+## v2.1.1 (2025-04-13)
+### Added
+
+## Changed
+* License updated to reflect current year
+
+### Fixed
+* The parameter for applying the DoD Adobe Acrobat Reader DC GPO was missing. Added the parameter and corresponding help.
+* The paths used for the DoD Adobe Acrobat Reader DC GPO were pointed to the Custom GPO directory. Corrected the paths to point to the DoD directory.
+
 ## v2.1.0 (2025-02-15)
 ### Added